각 리스트의 각 수정 Modify 화면 3개를 하나로 사용하게 만듬

+++ src/main/java/com/hongsi/purchshop/mapper/PurchOrderMapper.java
@@ -9,7 +9,7 @@ public interface PurchOrderMapper {
 
 	int insertPurchshopOrder(PurchOrderVO vo);
 	
-	int deleteOrderInfoByCno(int cno);
+	int deleteOrderInfoByCno(long cno);
 	
 	int updateOrderInfoByCno(PurchOrderVO vo);
